Refactor marker functionality into common base class. Track progress via a state callback on the marker.
